MAINT: Move the `flagsobj` annotations to `np.core.multiarray`
